requirements.txt
Django==1.9.5
python-dateutil==2.5.3
six==1.10.0
